WIP: refactor. ddd
