+++ src/domains/github_database_entry.ts
@@ -1,0 +1,9 @@
+interface DatabaseEntry {
+}
+
+export interface GithubDatabaseEntry extends DatabaseEntry {
+  owner: string;
+  repo: string;
+}
+
+export type GithubDatabaseEntries = Array<GithubDatabaseEntry>
