+++ src/services/github_service.ts
@@ -1,0 +1,13 @@
+import { Repository, Repositories } from "../domains/repository.ts";
+
+class GithubService {
+  constructor() {
+  }
+
+  public getGithubRepositories(): Repositories {
+    // TODO
+  }
+}
+
+// singleton
+export default new GithubService() as GithubService;
