+++ src/tsv_file_creator.ts
@@ -1,4 +1,4 @@
-import { Repository } from "./src/Repository.ts";
+import { Repository } from "./repositories/repository.ts";
 
 export async function generateTsvFile(result: Repository[]) {
   const encoder = new TextEncoder();
