+++ src/utils/unique.ts
@@ -1,4 +1,4 @@
-import { Repository } from "./Repository.ts";
+import { Repository } from "../domains/repository.ts";
 
 export function unique(repositories: Repository[]): Repository[] {
   return [...new Map(repositories.map((o) => [o.html_url, o])).values()];
